# src/agent/Hercules.WebApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Hercules.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/lifecycle/inventory": {
      "get": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecycleInventory",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/lifecycle/agent/{agentId}/start": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecycleAgentStart",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/agent/{agentId}/stop": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecycleAgentStop",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/agent/{agentId}/drain": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecycleAgentDrain",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/agent/{agentId}/decommission": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecycleAgentDecommission",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/agent/{agentId}/rollback": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecycleAgentRollback",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/packages/{packageId}/update": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecyclePackageUpdate",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/packages/{packageId}/canary": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecyclePackageCanary",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/packages/{packageId}/promote": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecyclePackagePromote",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteCanaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lifecycle/packages/{packageId}/rollback": {
      "post": {
        "tags": [
          "Lifecycle"
        ],
        "operationId": "LifecyclePackageRollback",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackPackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/escalations/pending": {
      "get": {
        "tags": [
          "Escalations"
        ],
        "operationId": "ListPendingEscalations",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minSeverity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/escalations/{id}": {
      "get": {
        "tags": [
          "Escalations"
        ],
        "operationId": "GetEscalation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/escalations/{id}/approve": {
      "post": {
        "tags": [
          "Escalations"
        ],
        "operationId": "ApproveEscalation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resolvedBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/escalations/{id}/deny": {
      "post": {
        "tags": [
          "Escalations"
        ],
        "operationId": "DenyEscalation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resolvedBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/escalations/batch-approve": {
      "post": {
        "tags": [
          "Escalations"
        ],
        "operationId": "BatchApproveEscalations",
        "parameters": [
          {
            "name": "resolvedBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchApproveRequest": {
        "required": [
          "ids"
        ],
        "type": "object",
        "properties": {
          "ids": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Request body for batch approve."
      },
...
      "CanaryRequest": {
        "required": [
          "targetAgentId"
        ],
        "type": "object",
        "properties": {
          "targetAgentId": {
            "type": "string"
          },
          "trafficPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 10
          }
        },
        "description": "Request to deploy a canary package."
      },
...
      "PromoteCanaryRequest": {
        "required": [
          "targetAgentId"
        ],
        "type": "object",
        "properties": {
          "targetAgentId": {
            "type": "string"
          }
        },
        "description": "Request to promote a canary to full deployment."
      },
...
      "RollbackPackageRequest": {
        "required": [
          "targetAgentId"
        ],
        "type": "object",
        "properties": {
          "targetAgentId": {
            "type": "string"
          }
        },
        "description": "Request to rollback a skill package."
      },
...
      "UpdatePackageRequest": {
        "required": [
          "targetAgentId"
        ],
        "type": "object",
        "properties": {
          "targetAgentId": {
            "type": "string"
          }
        },
        "description": "Request to update a skill package on target agent(s)."
      },
...
    }
  },
...
}